Refactor ImGuiZepConsole and Editor into zep project as standalone components
