+++ contrib/zep/zep.vcxproj
@@ -115,6 +115,8 @@
     </ClCompile>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="..\..\src\zep\ImGuiZepConsole.cpp" />
+    <ClCompile Include="..\..\src\zep\ImGuiZepEditor.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
@@ -181,6 +183,8 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="..\..\src\zep\ImGuiZepConsole.h" />
+    <ClInclude Include="..\..\src\zep\ImGuiZepEditor.h" />
     <ClInclude Include="include\zep.h" />
     <ClInclude Include="include\zep\buffer.h" />
     <ClInclude Include="include\zep\commands.h" />
