+++ contrib/zep/zep.vcxproj.filters
@@ -22,8 +22,14 @@
     <Filter Include="Header Files\zep">
       <UniqueIdentifier>{1ff7a728-1607-490a-95cf-28759e3a6653}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Header Files\imgui">
+      <UniqueIdentifier>{bb40859b-6c6c-4ba9-af38-915a4fc16091}</UniqueIdentifier>
+    </Filter>
     <Filter Include="Source Files\zep">
       <UniqueIdentifier>{c99b09a6-5a39-4da7-bf89-cdf1e24ec7e4}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Source Files\imgui">
+      <UniqueIdentifier>{90aa2191-5b47-408c-a311-0eca64b7535e}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
@@ -114,8 +120,14 @@
     <ClCompile Include="src\window.cpp">
       <Filter>Source Files\zep</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\zep\ImGuiZepEditor.cpp">
+      <Filter>Source Files\imgui</Filter>
+    </ClCompile>
     <ClCompile Include="pch.cpp">
       <Filter>Source Files\zep</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\zep\ImGuiZepConsole.cpp">
+      <Filter>Source Files\imgui</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -266,8 +278,14 @@
     <ClInclude Include="include\zep\zep_config.h">
       <Filter>Header Files\zep</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\zep\ImGuiZepEditor.h">
+      <Filter>Header Files\imgui</Filter>
+    </ClInclude>
     <ClInclude Include="pch.h">
       <Filter>Header Files\zep</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\zep\ImGuiZepConsole.h">
+      <Filter>Header Files\imgui</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
