+++ src/main/MQ2ImGuiConsole.cpp
@@ -16,16 +16,16 @@
 
 #include "MQ2DeveloperTools.h"
 #include "MQ2ImGuiTools.h"
-#include "MQ2Utilities.h"
-#include "ImGuiZepEditor.h"
 #include "ImGuiManager.h"
+#include "zep/ImGuiZepEditor.h"
+#include "zep/ImGuiZepConsole.h"
+#include "zep.h"
 
 #include "imgui/ImGuiTreePanelWindow.h"
 #include "mq/imgui/ConsoleWidget.h"
 
 #include <imgui/imgui_internal.h>
 
-#include "zep.h"
 #include <optional>
 #include "sqlite3.h"
 
@@ -44,13 +44,9 @@ bool gbAutoDockspaceViewport = false;
 bool gbAutoDockspacePreserveRatio = false;
 
 
-static const ImU32 s_defaultColor = Zep::ZepColor(240, 240, 240, 255);
+static constexpr MQColor s_defaultColor = MQColor(240, 240, 240, 255);
 static ImGuiID s_dockspaceId = 0;
 static ImGuiID s_dockspaceTopSegmentId = 0;
-
-// Some plain default colors
-static const ImU32 s_defaultLinkColor = Zep::ZepColor(0, 128, 255);
-static const ImU32 s_defaultLinkColorHover = Zep::ZepColor(255, 255, 128);
 
 // Some default color constants to patch eq style links
 static const ImU32 s_linkHoverColorDefault = Zep::ZepColor(0, 0, 255);
@@ -75,8 +71,8 @@ static bool s_resetConsolePosition = false;
 static bool s_setFocus = false;
 static bool s_consolePersistentCommandHistory = false;
 
-class ImGuiConsole;
-ImGuiConsole* gImGuiConsole = nullptr;
+class MQConsole;
+MQConsole* gMQConsole = nullptr;
 
 static void Strtrim(char* str)
 {
@@ -86,580 +82,41 @@ static void Strtrim(char* str)
 	*str_end = 0;
 }
 
-unsigned int str_to_hex(char const* p, char const* e) noexcept
-{
-	unsigned int result = 0;
-	while (p != e)
-	{
-		result *= 16;
-		if ('0' <= *p && *p <= '9') { result += *p - '0'; p++; continue; }
-		if ('A' <= *p && *p <= 'F') { result += *p + 10 - 'A'; p++; continue; }
-		if ('a' <= *p && *p <= 'f') { result += *p + 10 - 'a'; p++; continue; }
-		return -1;
-	}
-
-	return result;
-}
-
-ImU32 str_to_col(std::string_view str)
-{
-	if (str.length() != 7 || str[0] != '#') { return 0; }
-
-	auto r = str_to_hex(str.data() + 1, str.data() + 3);
-	auto g = str_to_hex(str.data() + 3, str.data() + 5);
-	auto b = str_to_hex(str.data() + 5, str.data() + 7);
-
-	return IM_COL32(r, g, b, 255);
-}
-
 static void MakeColorGradient(float frequency1, float frequency2, float frequency3,
 	float phase1, float phase2, float phase3,
 	float center = 128, float width = 127, int length = 50);
-
-static std::pair<std::string_view, ImU32> ParseColorTags(std::string_view line, std::vector<ImU32>& colorStack,
-	ImU32 defaultColor)
-{
-	size_t length = line.length();
-	const char* pos = line.data();
-	const char* end = line.data() + length;
-
-	ImU32 color = defaultColor;
-
-	// skip over the \a
-	if (*(pos++) != '\a')
-		return { {}, color };
-
-	if (pos == end) return { {}, color };
-
-	bool dark = false;
-
-	// clear
-	if (*pos == 'x')
-	{
-		pos++;
-
-		if (!colorStack.empty())
-		{
-			colorStack.pop_back();
-		}
-
-		if (!colorStack.empty())
-		{
-			color = colorStack.back();
-		}
-		else
-		{
-			color = defaultColor;
-		}
-
-		return { std::string_view{ pos, (size_t)(end - pos) }, color };
-	}
-
-	// custom color
-	if (*pos == '#')
-	{
-		// we need 7 to do anything (6 for hex code and 1 for #)
-		if (end - pos < 7) return { {}, color };
-		std::string_view colorCode{ pos, 7 };
-
-		// convert hex to color
-		color = str_to_col(colorCode);
-		colorStack.push_back(color);
-
-		pos += 7;
-		return { std::string_view{ pos, (size_t)(end - pos) }, color };
-	}
-
-	// darken
-	if (*pos == '-')
-	{
-		dark = true;
-		pos++;
-
-		if (pos == end) return { {}, color };
-	}
-
-	switch (*pos)
-	{
-	case 'y': // yellow (green/red)
-		if (dark)
-			color = 0xff009999;
-		else
-			color = 0xff00ffff;
-		break;
-	case 'o': // orange (green/red)
-		if (dark)
-			color = 0xff006699;
-		else
-			color = 0xff0099ff;
-		break;
-	case 'g': // green   (green)
-		if (dark)
-			color = 0xff009900;
-		else
-			color = 0xff00ff00;
-		break;
-	case 'u': // blue   (blue)
-		if (dark)
-			color = 0xff990000;
-		else
-			color = 0xffff0000;
-		break;
-	case 'r': // red     (red)
-		if (dark)
-			color = 0xff000099;
-		else
-			color = 0xff0000ff;
-		break;
-	case 't': // teal (blue/green)
-		if (dark)
-			color = 0xff999900;
-		else
-			color = 0xffffff00;
-		break;
-	case 'b': // black   (none)
-		color = 0xff000000;
-		break;
-	case 'm': // magenta (blue/red)
-		if (dark)
-			color = 0xff990099;
-		else
-			color = 0xffff00ff;
-		break;
-	case 'p': // purple (blue/red)
-		if (dark)
-			color = 0xff990066;
-		else
-			color = 0xffff0099;
-		break;
-	case 'w': // white   (all)
-		if (dark)
-			color = 0xff999999;
-		else
-			color = 0xffffffff;
-		break;
-	}
-	pos++;
-	colorStack.push_back(color);
-
-	return { { pos, (size_t)(end - pos) }, color };
-}
 
 //============================================================================
 
 #pragma region Zep Integration
 
-// This theme extends the default theme to support arbitrary coloring via extended
-// UserColor values of the ThemeColor enum.
-class ZepConsoleTheme : public Zep::ZepTheme
-{
-	static inline constexpr int UserColorStart = static_cast<int>(Zep::ThemeColor::UniqueColorLast) + 1;
-
-public:
-	ZepConsoleTheme()
-	{
-	}
-
-	virtual Zep::ZepColor GetColor(Zep::ThemeColor themeColor) const override
-	{
-		Zep::ZepColor finalColor;
-		
-		if (static_cast<int>(themeColor) >= UserColorStart)
-			finalColor = m_userColors[static_cast<size_t>(themeColor) - UserColorStart];
-		else if (themeColor == Zep::ThemeColor::Background)
-			finalColor = Zep::ZepColor(0, 0, 0, 0);
-		else if (themeColor == Zep::ThemeColor::VisualSelectBackground)
-			finalColor = Zep::ZepColor(66, 150, 249, 89);
-		else
-			finalColor = Zep::ZepTheme::GetColor(themeColor);
-		
-		// apply user alpha
-		finalColor.a = static_cast<uint8_t>(finalColor.a * m_opacity);
-
-		return finalColor;
-	}
-
-	// This overrides GetUniqueColor to treat any value over UniqueColorLast as a color value.
-	// It will be inserted into the user color map and the index + UserColorStart will be returned.
-	Zep::ThemeColor GetUserColor(Zep::ZepColor color)
-	{
-		uint32_t id = color.ABGR;
-		if (id < static_cast<int>(Zep::ThemeColor::UniqueColorLast) - static_cast<int>(Zep::ThemeColor::UniqueColor0))
-			return Zep::ZepTheme::GetUniqueColor(id);
-
-		auto iter = std::find(std::begin(m_userColors), std::end(m_userColors), color);
-		if (iter == std::end(m_userColors))
-		{
-			m_userColors.emplace_back(id);
-			return static_cast<Zep::ThemeColor>(m_userColors.size() - 1 + UserColorStart);
-		}
-
-		int index = static_cast<int>(std::distance(std::begin(m_userColors), iter));
-		return static_cast<Zep::ThemeColor>(index + UserColorStart);
-	}
-
-	float GetOpacity() { return m_opacity; }
-	void SetOpacity(const float opacity) { m_opacity = opacity; }
-
-private:
-	std::vector<Zep::ZepColor> m_userColors;
-	float m_opacity = 1.0f;
-};
-
 //----------------------------------------------------------------------------
 
-enum class ZepAttributeType
-{
-	Color,
-	Hyperlink,
-};
-
-struct ZepAttribute
-{
-	struct ColorAttributeData
-	{
-		uint32_t color;
-	};
-
-	struct HyperlinkAttributeData
-	{
-		std::string linkData;
-		uint32_t color = s_defaultLinkColor;
-		uint32_t hoverColor = s_defaultLinkColorHover;
-	};
-
-	using ZepAttributeData = std::variant<ColorAttributeData, HyperlinkAttributeData>;
-
-	ZepAttributeType type;
-	ZepAttributeData data;
-
-	ZepAttribute(ZepAttributeType type, ZepAttributeData data)
-		: type(type)
-		, data(std::move(data))
-	{
-	}
-
-	ZepAttribute() = default;
-};
-
-struct ZepTextAttribute
-{
-	int startIndex;
-	int endIndex;
-	ZepAttribute attribute;
-};
-
-struct ZepBufferAttribute
-{
-	Zep::GlyphIterator start;
-	int startIndex;
-	int endIndex;
-	ZepAttribute attribute;
-};
-
-using ZepAttributes = std::vector<ZepBufferAttribute>;
-
-constexpr const Zep::Msg UserEvent_HyperlinkLeftClick = static_cast<Zep::Msg>(static_cast<int>(Zep::Msg::UserEvent) + 1);
-constexpr const Zep::Msg UserEvent_HyperlinkRightClick = static_cast<Zep::Msg>(static_cast<int>(Zep::Msg::UserEvent) + 2);
-
-// This custom syntax makes use of ranged-based "attributes" that annotate the text to produce
-// colorization and hyperlinks for the editor.
-class ZepConsoleSyntax : public Zep::ZepSyntax
+class MQConsoleDelegate : public ImGuiZepConsoleDelegate
 {
 public:
-	ZepConsoleSyntax(Zep::ZepBuffer& buffer, const std::shared_ptr<ZepConsoleTheme>& theme, Zep::ZepWindow* window)
-		: Zep::ZepSyntax(buffer)
-		, m_theme(theme)
+	virtual void OnHyperlinkClicked(ImGuiZepConsole* console, Zep::ZepMouseButton button, uint32_t modifiers,
+		const std::string& hyperlinkData, int hyperlinkId) override
 	{
-		SetWindow(window);
-	}
-
-	void SetWindow(Zep::ZepWindow* window)
-	{
-		onMouseCursorChanged = window->sigMouseCursorChanged.connect(
-			[=](Zep::ZepBuffer& buffer, const Zep::GlyphIterator& mousePos)
-			{
-				if (&buffer != &m_buffer)
-					return;
-				UpdateMouseCursor(mousePos);
-			});
-	}
-
-	struct SyntaxData
-	{
-		Zep::ThemeColor foreground = Zep::ThemeColor::Normal;
-		uint32_t hyperlinkId = 0;
-	};
-
-	Zep::SyntaxResult GetSyntaxAt(const Zep::GlyphIterator& offset) const
-	{
-		Zep::SyntaxResult result{};
-
-		if (m_latestPosition < offset.Index() || (long)m_syntax.size() <= offset.Index())
+		if (starts_with(hyperlinkData, "testlink:"))
 		{
-			return result;
-		}
+			std::string text = fmt::format("Clicked hyperlink: {}\n", std::string_view{ hyperlinkData }.substr(9));
 
-		const SyntaxData& syntaxData = m_syntax[offset.Index()];
-
-		if (syntaxData.hyperlinkId)
-		{
-			result.foreground = GetHyperlinkColor(syntaxData.hyperlinkId, m_hoveredHyperlink == syntaxData.hyperlinkId);
+			console->AppendText(text, MQColor(255, 255, 0), true);
 		}
 		else
 		{
-			result.foreground = syntaxData.foreground;
-		}
+			TextTagInfo tagInfo = ExtractLink(hyperlinkData);
 
-		return result;
-	}
-
-	Zep::ThemeColor GetHyperlinkColor(uint32_t linkId, bool hovered) const
-	{
-		// Get the color tied to this hyperlink.
-		auto iter = m_hyperlinkData.find(linkId);
-		if (iter != m_hyperlinkData.end())
-		{
-			return m_theme->GetUserColor(hovered ? iter->second.hoverColor : iter->second.color);
-		}
-
-		return m_theme->GetUserColor(hovered ? s_defaultLinkColorHover : s_defaultLinkColor);
-	}
-
-	void Notify(std::shared_ptr<Zep::ZepMessage> spMsg)
-	{
-		if (spMsg->messageId == Zep::Msg::Buffer)
-		{
-			auto spBufferMsg = std::static_pointer_cast<Zep::BufferMessage>(spMsg);
-			if (spBufferMsg->pBuffer != &m_buffer)
+			if (!ExecuteTextLink(tagInfo))
 			{
-				return;
-			}
-
-			if (spBufferMsg->type == Zep::BufferMessageType::PreBufferChange)
-			{
-			}
-			else if (spBufferMsg->type == Zep::BufferMessageType::TextDeleted)
-			{
-				// Remove any hyperlinks in deleted text.
-				for (int pos = spBufferMsg->startLocation.Index(); pos != spBufferMsg->endLocation.Index(); ++pos)
-				{
-					if (m_syntax[pos].hyperlinkId != 0)
-					{
-						RemoveHyperlink(m_syntax[pos].hyperlinkId);
-					}
-				}
-
-				m_syntax.erase(m_syntax.begin() + spBufferMsg->startLocation.Index(),
-					m_syntax.begin() + spBufferMsg->endLocation.Index());
-
-				m_latestPosition -= (spBufferMsg->endLocation.Index() - spBufferMsg->endLocation.Index());
-			}
-			else if (spBufferMsg->type == Zep::BufferMessageType::TextAdded
-				|| spBufferMsg->type == Zep::BufferMessageType::Loaded)
-			{
-				auto dist = Zep::ByteDistance(spBufferMsg->startLocation, spBufferMsg->endLocation);
-				m_syntax.insert(m_syntax.begin() + spBufferMsg->startLocation.Index(), dist, SyntaxData{});
-
-				// Fill in syntax data from attributes.
-				for (const ZepBufferAttribute& attribute : m_pendingAttributes)
-				{
-					if (attribute.attribute.type == ZepAttributeType::Color)
-					{
-						uint32_t color = std::get<(int)ZepAttributeType::Color>(attribute.attribute.data).color;
-						Zep::ThemeColor themeColor = m_theme->GetUserColor(color);
-
-						Zep::GlyphRange range = { attribute.start + attribute.startIndex, attribute.start + attribute.endIndex };
-						for (Zep::GlyphIterator iter = range.first; iter != range.second; iter++)
-						{
-							m_syntax[iter.Index()].foreground = themeColor;
-						}
-					}
-
-					if (attribute.attribute.type == ZepAttributeType::Hyperlink)
-					{
-						auto& hyperlinkData = std::get<(int)ZepAttributeType::Hyperlink>(attribute.attribute.data);
-						uint32_t hyperlinkId = MakeHyperlink(hyperlinkData);
-
-						Zep::GlyphRange range = { attribute.start + attribute.startIndex, attribute.start + attribute.endIndex };
-						for (Zep::GlyphIterator iter = range.first; iter != range.second; iter++)
-						{
-							m_syntax[iter.Index()].hyperlinkId = hyperlinkId;
-						}
-					}
-				}
-				m_pendingAttributes.clear();
-				m_latestPosition += (spBufferMsg->endLocation.Index() - spBufferMsg->startLocation.Index());
-			}
-			else if (spBufferMsg->type == Zep::BufferMessageType::TextChanged)
-			{
-				auto dist = Zep::ByteDistance(spBufferMsg->startLocation, spBufferMsg->endLocation);
-
-				// Just clear syntax data. Changed text has no new attributes.
-				std::fill_n(m_syntax.begin() + spBufferMsg->startLocation.Index(), dist, SyntaxData{});
+				console->AppendText(fmt::format("Clicked link: {}\n", hyperlinkData));
 			}
 		}
 	}
 
-	void DispatchMouseEvent(std::shared_ptr<Zep::ZepMessage> message)
-	{
-		if (message->messageId == Zep::Msg::MouseDown
-			&& m_hoveredHyperlink != 0)
-		{
-			if (message->button == Zep::ZepMouseButton::Left)
-			{
-				auto hyperlinkMsg = std::make_shared<Zep::ZepMessage>(UserEvent_HyperlinkLeftClick, m_hyperlinkData[m_hoveredHyperlink].linkData);
-				GetEditor().Broadcast(hyperlinkMsg);
-
-				message->handled = true;
-			}
-		}
-		else if (message->messageId == Zep::Msg::MouseMove
-			&& message->button == Zep::ZepMouseButton::Left)
-		{
-			// Don't drag from links.
-			message->handled = true;
-		}
-	}
-
-	void AddAttribute(const Zep::GlyphIterator& position, ZepTextAttribute attr)
-	{
-		ZepBufferAttribute& buffAttr = m_pendingAttributes.emplace_back();
-		buffAttr.start = position;
-		buffAttr.startIndex = attr.startIndex;
-		buffAttr.endIndex = attr.endIndex;
-		buffAttr.attribute = std::move(attr.attribute);
-	}
-
-	uint32_t MakeHyperlink(const ZepAttribute::HyperlinkAttributeData& hyperlinkData)
-	{
-		uint32_t hyperlinkId = m_nextHyperlinkId++;
-		m_hyperlinkData[hyperlinkId] = hyperlinkData;
-		return hyperlinkId;
-	}
-
-	void RemoveHyperlink(uint32_t hyperlinkId)
-	{
-		m_hyperlinkData.erase(hyperlinkId);
-		m_hoveredHyperlink = 0;
-	}
-
-	void UpdateMouseCursor(const Zep::GlyphIterator& offset)
-	{
-		m_hoveredHyperlink = 0;
-
-		if (m_latestPosition < offset.Index() || (long)m_syntax.size() <= offset.Index())
-		{
-			return;
-		}
-
-		if (offset.Valid())
-		{
-			m_hoveredHyperlink = m_syntax[offset.Index()].hyperlinkId;
-		}
-	}
-
-private:
-	std::vector<SyntaxData> m_syntax;
-	std::shared_ptr<ZepConsoleTheme> m_theme;
-	std::vector<ZepBufferAttribute> m_pendingAttributes;
-	uint32_t m_nextHyperlinkId = 1;
-	std::map<uint32_t, ZepAttribute::HyperlinkAttributeData> m_hyperlinkData;
-	uint32_t m_hoveredHyperlink = 0;
-	Zep::scoped_connection onMouseCursorChanged;
-	int m_latestPosition = 0;
-};
-
-//----------------------------------------------------------------------------
-
-class ZepMode_ImGuiConsole : public Zep::ZepMode_Standard
-{
-public:
-	using ZepMode_Standard::ZepMode_Standard;
-
-	virtual Zep::CursorType GetCursorType() const override { return Zep::CursorType::None; }
-
-	static const char* StaticName()
-	{
-		return "Console";
-	}
-
-	virtual const char* Name() const override
-	{
-		return StaticName();
-	}
-};
-
-//----------------------------------------------------------------------------
-
-// This is the imgui container for the Zep component.
-struct ImGuiZepConsole : public mq::imgui::ConsoleWidget, public mq::imgui::ImGuiZepEditor
-{
-	Zep::ZepBuffer* m_buffer = nullptr;
-	Zep::ZepWindow* m_window = nullptr;
-	bool m_deferredCursorToEnd = false;
-	std::shared_ptr<ZepConsoleTheme> m_theme;
-	std::shared_ptr<ZepConsoleSyntax> m_syntax;
-	int m_maxBufferLines = 10000;
-	bool m_autoScroll = true;
-	std::string m_id;
-
-	ImGuiZepConsole(std::string_view id)
-		: m_id(std::string(id))
-	{
-		GetEditor().RegisterGlobalMode(std::make_shared<ZepMode_ImGuiConsole>(GetEditor()));
-		GetEditor().SetGlobalMode(ZepMode_ImGuiConsole::StaticName());
-
-		m_window = GetEditor().GetActiveTabWindow()->GetActiveWindow();
-		m_theme = std::make_shared<ZepConsoleTheme>();
-		GetEditor().SetTheme(m_theme);
-
-		GetEditor().GetConfig().style = Zep::EditorStyle::Minimal;
-		m_window->SetWindowFlags(Zep::WindowFlags::WrapText);
-
-		GetEditor().RegisterSyntaxFactory(
-			{ "Console" },
-			Zep::SyntaxProvider{ "Console", Zep::tSyntaxFactory([this](Zep::ZepBuffer* pBuffer) {
-				return std::make_shared<ZepConsoleSyntax>(*pBuffer, m_theme, m_window);
-			})
-		});
-
-		m_buffer = GetEditor().InitWithText("Console", "");
-		m_buffer->SetTheme(m_theme);
-		m_window->SetBufferCursor(m_buffer->End());
-		m_window->ToggleFlag(Zep::WindowFlags::HideTrailingNewline);
-		m_buffer->SetFileFlags(Zep::FileFlags::ReadOnly | Zep::FileFlags::CrudeUtf8Vaidate);
-	}
-
-	void Clear() override
-	{
-		m_buffer->Clear();
-	}
-
-	Zep::ZepWindow* GetWindow() const { return m_window; }
-	Zep::ZepBuffer* GetBuffer() const { return m_buffer; }
-
-	Zep::GlyphIterator InsertText(Zep::GlyphIterator position, std::string_view text, ImU32 color = -1)
-	{
-		if (color != -1)
-		{
-			ZepTextAttribute attribute;
-			attribute.startIndex = 0;
-			attribute.endIndex = static_cast<int>(text.length());
-			attribute.attribute.type = ZepAttributeType::Color;
-			attribute.attribute.data = ZepAttribute::ColorAttributeData{ color };
-
-			ZepConsoleSyntax* syntax = static_cast<ZepConsoleSyntax*>(m_buffer->GetSyntax());
-			syntax->AddAttribute(position, std::move(attribute));
-		}
-
-		Zep::ChangeRecord changeRecord;
-		m_buffer->Insert(position, text, changeRecord);
-
-		return position.Move(static_cast<long>(text.length()));
-	}
-
-	void InsertFormattedText(Zep::GlyphIterator position, std::string_view text, ImU32 color)
+	virtual bool OnInsertFormattedText(ImGuiZepConsole* console, Zep::GlyphIterator position, std::string_view text,
+		ImU32 color) override
 	{
 		// Parse hyperlink data
 		static TextTagInfo textTagInfo[MAX_EXTRACT_LINKS];
@@ -678,12 +135,12 @@ struct ImGuiZepConsole : public mq::imgui::ConsoleWidget, public mq::imgui::ImGu
 				std::string_view curSeg = text.substr(segPos, tagInfo.link.data() - text.data() - segPos);
 				if (!curSeg.empty())
 				{
-					position = InsertText(position, curSeg, color);
+					position = console->InsertText(position, curSeg, color);
 					segPos += curSeg.length();
 				}
 
 				// Insert hyperlink.
-				InsertHyperlink(position, tagInfo);
+				InsertHyperlink(console, position, tagInfo);
 				position = position.Move(static_cast<long>(tagInfo.text.length()));
 				segPos = tagInfo.link.data() - text.data() + tagInfo.link.size();
 			}
@@ -692,16 +149,17 @@ struct ImGuiZepConsole : public mq::imgui::ConsoleWidget, public mq::imgui::ImGu
 			std::string_view endSeg = text.substr(segPos);
 			if (!endSeg.empty())
 			{
-				position = InsertText(position, endSeg, color);
+				position = console->InsertText(position, endSeg, color);
 			}
+
+			return true;
 		}
-		else
-		{
-			InsertText(position, text, color);
-		}
+
+		return false;
 	}
 
-	void InsertHyperlink(Zep::GlyphIterator position, const TextTagInfo& tagInfo)
+private:
+	void InsertHyperlink(ImGuiZepConsole* console, Zep::GlyphIterator position, const TextTagInfo& tagInfo)
 	{
 		uint32_t color = s_linkColorDefault;
 		uint32_t hoverColor = s_linkHoverColorDefault;
@@ -718,7 +176,7 @@ struct ImGuiZepConsole : public mq::imgui::ConsoleWidget, public mq::imgui::ImGu
 			break;
 		case ETAG_SPAM:
 			color = s_linkColorSpam;
-			color = s_linkHoverColorSpam;
+			hoverColor = s_linkHoverColorSpam;
 			break;
 		case ETAG_ACHIEVEMENT:
 			color = GetColorForChatColor(s_userColorAchievementLink).ToABGR();
@@ -738,170 +196,7 @@ struct ImGuiZepConsole : public mq::imgui::ConsoleWidget, public mq::imgui::ImGu
 			break;
 		}
 
-		InsertHyperlink(position, tagInfo.text, std::string(tagInfo.link), color, hoverColor);
-	}
-
-	void InsertHyperlink(Zep::GlyphIterator position, std::string_view text, std::string hyperlinkData, uint32_t color, uint32_t hoverColor)
-	{
-		ZepTextAttribute attribute;
-		attribute.startIndex = 0;
-		attribute.endIndex = static_cast<int>(text.length());
-		attribute.attribute.type = ZepAttributeType::Hyperlink;
-		attribute.attribute.data = ZepAttribute::HyperlinkAttributeData{ std::move(hyperlinkData), color, hoverColor };
-
-		ZepConsoleSyntax* syntax = static_cast<ZepConsoleSyntax*>(m_buffer->GetSyntax());
-		syntax->AddAttribute(position, std::move(attribute));
-
-		Zep::ChangeRecord changeRecord;
-		m_buffer->Insert(position, text, changeRecord);
-	}
-
-	// This accepts color in ABGR.
-	void AppendFormattedText(std::string_view text, uint32_t defaultColor = s_defaultColor, bool newline = false)
-	{
-		Zep::GlyphIterator cursor = m_window->GetBufferCursor();
-		bool cursorAtEnd = m_window->IsAtBottom();
-
-		std::string_view lineView = text;
-		ImU32 currentColor = defaultColor;
-
-		std::vector<ImU32> colorStack;
-
-		while (!lineView.empty())
-		{
-			auto colorPos = lineView.find("\a");
-
-			// this is everything before the color code.
-			auto beforeColor = lineView.substr(0, colorPos);
-			if (!beforeColor.empty())
-			{
-				// no color codes, write out with current color
-				InsertFormattedText(m_buffer->End(), beforeColor, currentColor);
-			}
-
-			// did we find a color?
-			if (colorPos == std::string_view::npos)
-				break;
-
-			lineView = lineView.substr(colorPos);
-
-			auto& [nextSegment, nextColor] = ParseColorTags(lineView, colorStack, defaultColor);
-			// Parse the color and get the next segment. We pass in the
-			// default color to handle \ax properly
-
-			if (nextSegment.empty())
-				break;
-
-			currentColor = nextColor;
-			lineView = nextSegment;
-		}
-
-		if (newline)
-			InsertText(m_buffer->End(), "\n");
-
-		PruneBuffer();
-
-		if (cursorAtEnd)
-		{
-			TriggerAutoScroll();
-		}
-	}
-
-	void AppendText(std::string_view text, MQColor defaultColor /* = DEFAULT_COLOR */, bool appendNewLine /* = false */) override
-	{
-		AppendFormattedText(text, defaultColor.ToImU32(), appendNewLine);
-	}
-
-	void TriggerAutoScroll()
-	{
-		if (m_autoScroll)
-		{
-			m_deferredCursorToEnd = true;
-		}
-	}
-
-	void ScrollToBottom() override
-	{
-		m_deferredCursorToEnd = true;
-	}
-
-	bool IsCursorAtEnd() const override
-	{
-		return m_window->IsAtBottom();
-	}
-
-	void PruneBuffer()
-	{
-		int lineCount = m_buffer->GetLineCount();
-		if (lineCount > m_maxBufferLines + 1)
-		{
-			int linesToDelete = lineCount - (m_maxBufferLines + 1);
-
-			Zep::ByteRange range;
-			if (m_buffer->GetLineOffsets(linesToDelete, range))
-			{
-				Zep::GlyphIterator end(m_buffer, range.first);
-
-				Zep::ChangeRecord changeRecord;
-				m_buffer->Delete(m_buffer->Begin(), end, changeRecord);
-			}
-		}
-	}
-
-	void Render(const ImVec2& displaySize = ImVec2()) override
-	{
-		if (m_deferredCursorToEnd)
-		{
-			m_deferredCursorToEnd = false;
-			m_window->ScrollToBottom();
-		}
-
-		ImGuiZepEditor::Render(m_id.c_str(), displaySize);
-	}
-
-	void Notify(std::shared_ptr<Zep::ZepMessage> message) override
-	{
-		if (message->messageId == UserEvent_HyperlinkLeftClick)
-		{
-			if (starts_with(message->str, "testlink:"))
-			{
-				std::string text = fmt::format("Clicked hyperlink: {}\n", std::string_view{ message->str }.substr(9));
-
-				AppendFormattedText(text, Zep::ZepColor(255, 255, 0));
-			}
-			else
-			{
-				TextTagInfo tagInfo = ExtractLink(message->str);
-
-				if (!ExecuteTextLink(tagInfo))
-				{
-					AppendFormattedText(fmt::format("Clicked link: {}\n", message->str));
-				}
-			}
-		}
-
-		ImGuiZepEditor::Notify(message);
-	}
-
-	bool GetAutoScroll() const override { return m_autoScroll; }
-
-	void SetAutoScroll(bool autoScroll) override
-	{
-		m_autoScroll = autoScroll;
-	}
-
-	int GetMaxBufferLines() const override { return m_maxBufferLines; }
-
-	void SetMaxBufferLines(int maxBufferLines) override
-	{
-		m_maxBufferLines = maxBufferLines;
-	}
-
-	float GetOpacity() const override { return m_theme->GetOpacity(); }
-
-	void SetOpacity(float opacity) override
-	{
-		m_theme->SetOpacity(opacity);
+		console->InsertHyperlink(position, tagInfo.text, std::string(tagInfo.link), color, hoverColor);
 	}
 };
 
@@ -911,7 +206,9 @@ namespace imgui
 {
 	std::shared_ptr<mq::imgui::ConsoleWidget> ConsoleWidget::Create(std::string_view id)
 	{
-		return std::make_shared<ImGuiZepConsole>(id);
+		auto console = std::make_shared<ImGuiZepConsole>(id);
+		console->SetDelegate(std::make_shared<MQConsoleDelegate>());
+		return console;
 	}
 }
 
@@ -1055,7 +352,7 @@ void AddEntryToDatabase(sqlite3* db, int process_id, const char* entry)
 
 #pragma region ImGui Console
 
-class ImGuiConsole
+class MQConsole
 {
 public:
 	char m_inputBuffer[2048];
@@ -1065,26 +362,26 @@ public:
 	int current_pid = GetCurrentProcessId();
 	int m_historyPos = -1;    // -1: new line, 0..History.Size-1 browsing history.
 	bool m_scrollToBottom = true;
-	std::unique_ptr<ImGuiZepConsole> m_zepEditor;
+	std::unique_ptr<ImGuiZepConsole> m_zepConsole;
 	bool m_localEcho = true;
 
-
-	ImGuiConsole()
+	MQConsole()
 	{
 		ZeroMemory(m_inputBuffer, lengthof(m_inputBuffer));
-		m_zepEditor = std::make_unique<ImGuiZepConsole>("##ZepConsole");
+		m_zepConsole = std::make_unique<ImGuiZepConsole>("##ZepConsole");
+		m_zepConsole->SetDelegate(std::make_shared<MQConsoleDelegate>());
 
 		m_localEcho = GetPrivateProfileBool("Console", "LocalEcho", m_localEcho, internal_paths::MQini);
 
-		bool autoScroll = GetPrivateProfileBool("Console", "AutoScroll", m_zepEditor->GetAutoScroll(), internal_paths::MQini);
-		m_zepEditor->SetAutoScroll(autoScroll);
+		bool autoScroll = GetPrivateProfileBool("Console", "AutoScroll", m_zepConsole->GetAutoScroll(), internal_paths::MQini);
+		m_zepConsole->SetAutoScroll(autoScroll);
 
-		int maxBufferLines = GetPrivateProfileInt("Console", "MaxBufferLines", m_zepEditor->GetMaxBufferLines(), internal_paths::MQini);
-		m_zepEditor->SetMaxBufferLines(maxBufferLines);
+		int maxBufferLines = GetPrivateProfileInt("Console", "MaxBufferLines", m_zepConsole->GetMaxBufferLines(), internal_paths::MQini);
+		m_zepConsole->SetMaxBufferLines(maxBufferLines);
 		m_history = InitConsoleDatabase(m_db, current_pid);
 	}
 
-	~ImGuiConsole()
+	~MQConsole()
 	{
 		ClearLog();
 		if (m_db != nullptr)
@@ -1095,7 +392,7 @@ public:
 
 	void ClearLog()
 	{
-		m_zepEditor->Clear();
+		m_zepConsole->Clear();
 	}
 
 	template <typename... Args>
@@ -1104,18 +401,18 @@ public:
 		fmt::basic_memory_buffer<char> buf;
 		fmt::format_to(fmt::appender(buf), fmt, args...);
 
-		m_zepEditor->AppendFormattedText(std::string_view(buf.data(), buf.size()), color, false);
+		m_zepConsole->AppendText(std::string_view(buf.data(), buf.size()), color, false);
 	}
 
 	template <typename... Args>
 	void AddLog(std::string_view fmt, const Args&... args)
 	{
-		AddLog(s_defaultColor, std::move(fmt), args...);
+		AddLog(s_defaultColor.ToImU32(), std::move(fmt), args...);
 	}
 
-	void AddWriteChatColorLog(const char* line, ImU32 defaultColor = s_defaultColor, bool newline = false)
+	void AddWriteChatColorLog(const char* line, MQColor defaultColor = s_defaultColor, bool newline = false)
 	{
-		m_zepEditor->AppendFormattedText(line, defaultColor, newline);
+		m_zepConsole->AppendText(line, defaultColor, newline);
 	}
 
 	void Draw(bool* pOpen)
@@ -1143,10 +440,10 @@ public:
 		{
 			if (ImGui::BeginMenu("Options"))
 			{
-				bool autoScroll = m_zepEditor->GetAutoScroll();
+				bool autoScroll = m_zepConsole->GetAutoScroll();
 				if (ImGui::MenuItem("Auto-scroll", nullptr, &autoScroll))
 				{
-					m_zepEditor->SetAutoScroll(autoScroll);
+					m_zepConsole->SetAutoScroll(autoScroll);
 					WritePrivateProfileBool("Console", "AutoScroll", autoScroll, internal_paths::MQini);
 				}
 
@@ -1184,7 +481,7 @@ public:
 						MakeColorGradient(.3f, .3f, .3f, 0, 2, 4);
 					}
 
-					if (m_zepEditor)
+					if (m_zepConsole)
 					{
 						if (ImGui::MenuItem("Hyperlink Test"))
 						{
@@ -1228,7 +525,7 @@ public:
 		ImVec2 contentSize = ImGui::GetContentRegionAvail();
 		contentSize.y -= footer_height_to_reserve;
 
-		m_zepEditor->Render(contentSize);
+		m_zepConsole->Render(contentSize);
 
 		// Command-line
 		ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, ImVec2(2, 4));
@@ -1246,7 +543,7 @@ public:
 
 		bool bTextEdit = ImGui::InputText("##Input", m_inputBuffer, IM_ARRAYSIZE(m_inputBuffer), textFlags,
 			[](ImGuiInputTextCallbackData* data)
-		{ return static_cast<ImGuiConsole*>(data->UserData)->TextEditCallback(data); }, this);
+		{ return static_cast<MQConsole*>(data->UserData)->TextEditCallback(data); }, this);
 
 		ImGui::PopFont();
 		ImGui::PopStyleColor();
@@ -1277,7 +574,7 @@ public:
 	void ExecCommand(const char* commandLine)
 	{
 		if (GetLocalEcho())
-			AddLog(Zep::ZepColor(128, 128, 128), "> {0}\n", commandLine);
+			AddLog(MQColor(128, 128, 128).ToImU32(), "> {0}\n", commandLine);
 
 		// Insert into history. First find match and delete it so i can be pushed to the back. This isn't
 		// trying to be smart or optimal.
@@ -1438,12 +735,12 @@ public:
 	void DoAchievementLinkTest()
 	{
 		std::string_view line = "You say to your guild, '\x12" "3TestToon^500010200^1^0^0^0^0^0^'Welcome to Crescent Reach (1+)\x12'";
-		m_zepEditor->AppendFormattedText(line, s_defaultColor, true);
+		m_zepConsole->AppendText(line, s_defaultColor, true);
 	}
 
 	void DoHyperlinkTest()
 	{
-		bool cursorAtEnd = m_zepEditor->GetWindow()->IsAtBottom();
+		bool cursorAtEnd = m_zepConsole->GetWindow()->IsAtBottom();
 
 		static int hyperlinkNum = 1;
 		std::string text = fmt::format("This is hyperlink {}", hyperlinkNum++);
@@ -1451,21 +748,21 @@ public:
 		ZepTextAttribute attribute;
 		attribute.startIndex = 0;
 		attribute.endIndex = static_cast<int>(text.length());
-		attribute.attribute.type = ZepAttributeType::Hyperlink;
+		attribute.attribute.type = ZepAttribute::Type::Hyperlink;
 		attribute.attribute.data = ZepAttribute::HyperlinkAttributeData{ fmt::format("testlink:{}'s data", text) };
 
 		// Append to end of buffer
-		Zep::GlyphIterator position = m_zepEditor->GetBuffer()->End();
+		Zep::GlyphIterator position = m_zepConsole->GetBuffer()->End();
 
-		ZepConsoleSyntax* syntax = static_cast<ZepConsoleSyntax*>(m_zepEditor->GetBuffer()->GetSyntax());
+		ImGuiZepConsoleSyntax* syntax = static_cast<ImGuiZepConsoleSyntax*>(m_zepConsole->GetBuffer()->GetSyntax());
 		syntax->AddAttribute(position, std::move(attribute));
 
 		Zep::ChangeRecord changeRecord;
-		m_zepEditor->GetBuffer()->Insert(position, text + "\n", changeRecord);
+		m_zepConsole->GetBuffer()->Insert(position, text + "\n", changeRecord);
 
 		if (cursorAtEnd)
 		{
-			m_zepEditor->TriggerAutoScroll();
+			m_zepConsole->TriggerAutoScroll();
 		}
 	}
 
@@ -1522,7 +819,7 @@ static void MakeColorGradient(float frequency1, float frequency2, float frequenc
 	float phase1, float phase2, float phase3,
 	float center, float width, int length)
 {
-	if (!gImGuiConsole)
+	if (!gMQConsole)
 		return;
 
 	char szBuffer[2048] = { 0 };
@@ -1644,7 +941,7 @@ void UpdateImGuiConsole()
 		}
 		s_resetConsolePosition = false;
 
-		gImGuiConsole->Draw(&s_consoleVisible);
+		gMQConsole->Draw(&s_consoleVisible);
 
 		if (shouldSetFocusToMainViewport)
 		{
@@ -1710,8 +1007,8 @@ void MQConsoleCommand(SPAWNINFO* pChar, char* Line)
 
 	if (ci_equals("clear", szCommand))
 	{
-		if (gImGuiConsole != nullptr)
-			gImGuiConsole->ClearLog();
+		if (gMQConsole != nullptr)
+			gMQConsole->ClearLog();
 
 		return;
 	}
@@ -1758,15 +1055,15 @@ static void ConsoleSettings()
 
 	ImGui::NewLine();
 
-	if (gImGuiConsole != nullptr)
+	if (gMQConsole != nullptr)
 	{
 		ImGui::Text("Maximum Number of Buffer Lines");
 
-		int maxBufferLines = gImGuiConsole->m_zepEditor->GetMaxBufferLines();
+		int maxBufferLines = gMQConsole->m_zepConsole->GetMaxBufferLines();
 		if (ImGui::InputInt("##BufferLineMaxEntry", &maxBufferLines))
 		{
 			WritePrivateProfileInt("Console", "MaxBufferLines", maxBufferLines, internal_paths::MQini);
-			gImGuiConsole->m_zepEditor->SetMaxBufferLines(maxBufferLines);
+			gMQConsole->m_zepConsole->SetMaxBufferLines(maxBufferLines);
 		}
 
 		ImGui::SameLine();
@@ -1797,14 +1094,14 @@ void InitializeImGuiConsole()
 
 	AddSettingsPanel("Console", ConsoleSettings);
 
-	gImGuiConsole = new ImGuiConsole();
+	gMQConsole = new MQConsole();
 	AddCommand("/mqconsole", MQConsoleCommand);
 }
 
 void ShutdownImGuiConsole()
 {
-	delete gImGuiConsole;
-	gImGuiConsole = nullptr;
+	delete gMQConsole;
+	gMQConsole = nullptr;
 
 	if (gbAutoDockspaceViewport)
 	{
@@ -1819,8 +1116,8 @@ DWORD ImGuiConsoleAddText(const char* line, DWORD color, DWORD filter)
 {
 	ImU32 col = GetColorForChatColor(color).ToABGR();
 
-	if (gImGuiConsole)
-		gImGuiConsole->AddWriteChatColorLog(line, col, true);
+	if (gMQConsole)
+		gMQConsole->AddWriteChatColorLog(line, col, true);
 
 	return 0;
 }
