+++ src/main/MQ2WindowInspector.cpp
@@ -15,11 +15,11 @@
 #include "pch.h"
 #include "MQ2DeveloperTools.h"
 
-#include "ImGuiZepEditor.h"
 
 #include "imgui/fonts/IconsFontAwesome.h"
 #include "imgui/imgui_internal.h"
 #include "imgui/ImGuiMemoryEditor.h"
+#include "zep/ImGuiZepEditor.h"
 #include "zep.h"
 
 #include "mq/imgui/ImGuiUtils.h"
